#38
- preset/isomorphic: модуль провайдеров разбит на отдельные модули по провайдерам (patch)
