+++ src/preset/isomorphic/providers/abort-controller.ts
@@ -1,0 +1,7 @@
+/**
+ * Провайдер AbortController.
+ * @return AbortController.
+ */
+export function provideAbortController(): AbortController {
+  return new AbortController();
+}
